sql-scripts/sql_monitor/sql-server-metrics/pyproject.toml
[project]
name = "sql-server-metrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.0",
    "opentelemetry-exporter-prometheus>=0.52b0",
    "opentelemetry-sdk>=1.31.0",
    "pymssql>=2.3.2",
]
